+++ src/main/java/edu/illinois/mitra/cyphyhouse/objects/Point.java
@@ -4,19 +4,19 @@ package edu.illinois.mitra.cyphyhouse.objects;
  * Created by SC on 11/11/16.
  */
 public class Point {
-    public int x,y;
-    public Point(int x, int y){
+    public double x,y;
+    public Point(double x, double y){
         set(x, y);
     }
     public Point(Point src){
         set(src.x, src.y);
     }
 
-    public final boolean equals(int x, int y){
+    public final boolean equals(double x, double y){
         return (this.x==x && this.y==y);
     }
 
-    public void set(int x, int y){
+    public void set(double x, double y){
         this.x = x;
         this.y = y;
     }
